Fix send bytom share log logic in sserver
